+++ src/lib/stats.c
@@ -144,6 +144,7 @@ void notcurses_stats_reset(notcurses* nc, ncstats* stats){
   stash->sprixelemissions += nc->stats.sprixelemissions;
   stash->sprixelelisions += nc->stats.sprixelelisions;
   stash->sprixelbytes += nc->stats.sprixelbytes;
+  stash->appsync_updates += nc->stats.appsync_updates;
 
   stash->fbbytes = nc->stats.fbbytes;
   stash->planes = nc->stats.planes;
@@ -214,10 +215,10 @@ void summarize_stats(notcurses* nc){
             (stats->bgelisions * 100.0) / (stats->bgemissions + stats->bgelisions));
     char totalbuf[BPREFIXSTRLEN + 1];
     qprefix(stats->sprixelbytes, 1, totalbuf, 1);
-    fprintf(stderr, "Sprixel emits:elides: %ju:%ju (%.2f%%) %sB\n",
+    fprintf(stderr, "Sprixel emits:elides: %ju:%ju (%.2f%%) %sB ASUs: %ju\n",
             stats->sprixelemissions, stats->sprixelelisions,
             (stats->sprixelemissions + stats->sprixelelisions) == 0 ? 0 :
             (stats->sprixelelisions * 100.0) / (stats->sprixelemissions + stats->sprixelelisions),
-            totalbuf);
+            totalbuf, stats->appsync_updates);
   }
 }